Move test service into test suite file.
