+++ middleware/grpc/grpc_suite_test.go
@@ -3,6 +3,11 @@
 package grpc_test
 
 import (
+	"context"
+	"errors"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/metadata"
+	"google.golang.org/grpc/status"
 	"net"
 	"testing"
 
@@ -10,7 +15,6 @@ import (
 	. "github.com/onsi/gomega"
 	"google.golang.org/grpc"
 
-	grpc_testing "github.com/openzipkin/zipkin-go/middleware/grpc/internal/testing"
 	"github.com/openzipkin/zipkin-go/model"
 	service "github.com/openzipkin/zipkin-go/proto/testing"
 )
@@ -28,7 +32,7 @@ var _ = BeforeSuite(func() {
 	Expect(lis, err).ToNot(BeNil(), "failed to listen to tcp port")
 
 	server = grpc.NewServer()
-	service.RegisterHelloServiceServer(server, &grpc_testing.TestHelloService{})
+	service.RegisterHelloServiceServer(server, &TestHelloService{})
 	go func() {
 		_ = server.Serve(lis)
 	}()
@@ -61,4 +65,29 @@ func (g *sequentialIdGenerator) TraceID() model.TraceID {
 	}
 	g.nextTraceId++
 	return id
+}
+
+type TestHelloService struct{}
+
+func (s *TestHelloService) Hello(ctx context.Context, req *service.HelloRequest) (*service.HelloResponse, error) {
+	if req.Payload == "fail" {
+		return nil, status.Error(codes.Aborted, "fail")
+	}
+
+	md, ok := metadata.FromIncomingContext(ctx)
+	if !ok {
+		return nil, errors.New("could not parse incoming metadata")
+	}
+
+	resp := &service.HelloResponse{
+		Payload:  "World",
+		Metadata: map[string]string{},
+	}
+
+	for k, _ := range md {
+		// Just append the first value for a key for simplicity since we don't use multi-value headers.
+		resp.GetMetadata()[k] = md[k][0]
+	}
+
+	return resp, nil
 }
